Manage socket

+++ src/main/java/speed/daemon/serverMessages/ServerMessage.java
@@ -1,5 +1,5 @@
 package speed.daemon.serverMessages;
 
 public interface ServerMessage {
-    int[] encode();
+    byte[] encode();
 }
